Using the RW try to verify valid storage namespacs, and printing the error messages that thrown in the procces
In the function ensureStorageNamespace there is a try to read and write
a dummy file, and in this try there is a verification of whether the
storage type in the create repo request is the same as the one in the configuration.
In this case, there is no need in separate verification.
I changed the error message to print the error thrown in the RW try.
Also, I changed the ensure function name to be less specific.